… src/manifold/src/face_op.cpp
… src/manifold/src/impl.h
… src/manifold/src/manifold.cpp