Use tbb for some optimizations directly:
1. Parallelize complex face triangulation. The approach with `std::async` will create a thread for every invocation, which is too much overhead, so tbb `thread_arena` is needed.
2. Parallelize `AddNewEdgeVerts`. This requires `concurrent_map` as we will be obtaining elements concurrently.

Combining these two optimizations, we can cut down the running time for `Samples.Sponge4` from ~4s to ~3s. When `MANIFOLD_DEBUG=on`, the running time for `Samples.Sponge4` is reduced from ~7s to ~3.7s. The main bottleneck for now should be simplification which takes about 1s.