@pca006132 2023-08-04
Maybe enqueue can cause a deadlock if itself is the sole worker thread... I guess I should use task_group instead.