@pca006132 2023-08-04
It seems that face reordering causes `SimplifyTopology` failure (I triggered both segfault and infinite loop). May be related to #518.