@pca006132 2023-08-04
thinking about it, it should be possible to parallelize this without reordering. will try it tmr