@pca006132 2023-08-05
Admittedly I am getting a bit crazy now, I was seeing so many optimization opportunities and can hardly sleep without implementing them.

Samples.Sponge4 now runs in 2600ms with tbb (it took 4000ms previously) and MANIFOLD_DEBUG=off, the entire test suite completes in 4.5s, and the perfTest with tbb:
```
nTri = 512, time = 0.00160684 sec
nTri = 2048, time = 0.00310806 sec
nTri = 8192, time = 0.00775844 sec
nTri = 32768, time = 0.0200173 sec
nTri = 131072, time = 0.029914 sec
nTri = 524288, time = 0.0974839 sec
nTri = 2097152, time = 0.434753 sec
nTri = 8388608, time = 1.94247 sec
```