@pca006132 2023-08-05
the segfault is weird, it doesn't seem to me that the collider update is doing anything that will cause segfault.