@pca006132 2023-08-06
Only not available on WASM. Yes, we can mark the old code as about to be removed.
I actually tried to compile tbb from source for WASM, it seems that it can work, but the performance is not great. I can look into it later.